Move user search to core
