+++ packages/backend-core/src/db/utils.ts
@@ -434,8 +434,8 @@ export const getPluginParams = (pluginId?: string | null, otherProps = {}) => {
   return getDocParams(DocumentType.PLUGIN, pluginId, otherProps)
 }
 
-export function pagination(
-  data: any[],
+export function pagination<T>(
+  data: T[],
   pageSize: number,
   {
     paginate,
@@ -444,7 +444,7 @@ export function pagination(
   }: {
     paginate: boolean
     property: string
-    getKey?: (doc: any) => string | undefined
+    getKey?: (doc: T) => string | undefined
   } = {
     paginate: true,
     property: "_id",
